Moved validation to about to submit instead of about to start
